// frontend/lib/api/generated/model/guiaRemisionListOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface GuiaRemisionListOut {
  guia_remision_id: number;
  numero_guia: string;
  proveedor_id: number;
  orden_compra_id: number;
  pedido_semanal_id: number;
  almacen_destino_id: number;
  fecha_entrega: string;
  estado: string;
  creado_en: string;
  proveedor_razon_social: string;
  almacen_destino_codigo: string;
  almacen_destino_nombre: string;
}